+++ src/systems/a320_systems/src/hydraulic/mod.rs
@@ -158,6 +158,7 @@ impl A320HydraulicCircuitFactory {
             Pressure::new::<psi>(Self::MIN_PRESS_EDP_SECTION_HI_HYST),
             true,
             false,
+            false,
             Pressure::new::<psi>(Self::HYDRAULIC_TARGET_PRESSURE_PSI),
         )
     }
@@ -175,6 +176,7 @@ impl A320HydraulicCircuitFactory {
             Pressure::new::<psi>(Self::MIN_PRESS_PRESSURISED_HI_HYST),
             Pressure::new::<psi>(Self::MIN_PRESS_BLUE_ELEC_PUMP_SECTION_LO_HYST),
             Pressure::new::<psi>(Self::MIN_PRESS_BLUE_ELEC_PUMP_SECTION_HI_HYST),
+            false,
             false,
             false,
             Pressure::new::<psi>(Self::HYDRAULIC_TARGET_PRESSURE_PSI),
@@ -196,6 +198,7 @@ impl A320HydraulicCircuitFactory {
             Pressure::new::<psi>(Self::MIN_PRESS_EDP_SECTION_HI_HYST),
             false,
             true,
+            false,
             Pressure::new::<psi>(Self::HYDRAULIC_TARGET_PRESSURE_PSI),
         )
     }
@@ -2045,116 +2048,117 @@ impl A320Hydraulic {
 
     fn update_green_actuators_volume(&mut self) {
         self.green_circuit
-            .update_actuator_volumes(&mut self.braking_circuit_norm);
+            .update_system_actuator_volumes(&mut self.braking_circuit_norm);
 
-        self.green_circuit
-            .update_actuator_volumes(self.left_aileron.actuator(AileronActuatorPosition::Inboard));
-        self.green_circuit.update_actuator_volumes(
+        self.green_circuit.update_system_actuator_volumes(
+            self.left_aileron.actuator(AileronActuatorPosition::Inboard),
+        );
+        self.green_circuit.update_system_actuator_volumes(
             self.right_aileron
                 .actuator(AileronActuatorPosition::Inboard),
         );
 
-        self.green_circuit.update_actuator_volumes(
+        self.green_circuit.update_system_actuator_volumes(
             self.left_elevator
                 .actuator(ElevatorActuatorPosition::Inboard),
         );
 
         self.green_circuit
-            .update_actuator_volumes(self.rudder.actuator(RudderActuatorPosition::Green));
+            .update_system_actuator_volumes(self.rudder.actuator(RudderActuatorPosition::Green));
 
         self.green_circuit
-            .update_actuator_volumes(self.flap_system.left_motor());
+            .update_system_actuator_volumes(self.flap_system.left_motor());
         self.green_circuit
-            .update_actuator_volumes(self.slat_system.right_motor());
+            .update_system_actuator_volumes(self.slat_system.right_motor());
 
         self.green_circuit
-            .update_actuator_volumes(self.left_spoilers.actuator(0));
+            .update_system_actuator_volumes(self.left_spoilers.actuator(0));
         self.green_circuit
-            .update_actuator_volumes(self.left_spoilers.actuator(4));
+            .update_system_actuator_volumes(self.left_spoilers.actuator(4));
 
         self.green_circuit
-            .update_actuator_volumes(self.right_spoilers.actuator(0));
+            .update_system_actuator_volumes(self.right_spoilers.actuator(0));
         self.green_circuit
-            .update_actuator_volumes(self.right_spoilers.actuator(4));
+            .update_system_actuator_volumes(self.right_spoilers.actuator(4));
 
         for actuator in self.gear_system.all_actuators() {
-            self.green_circuit.update_actuator_volumes(actuator);
+            self.green_circuit.update_system_actuator_volumes(actuator);
         }
 
         self.green_circuit
-            .update_actuator_volumes(self.trim_assembly.left_motor());
+            .update_system_actuator_volumes(self.trim_assembly.left_motor());
     }
 
     fn update_yellow_actuators_volume(&mut self) {
         self.yellow_circuit
-            .update_actuator_volumes(&mut self.braking_circuit_altn);
+            .update_system_actuator_volumes(&mut self.braking_circuit_altn);
 
         self.yellow_circuit
-            .update_actuator_volumes(self.flap_system.right_motor());
+            .update_system_actuator_volumes(self.flap_system.right_motor());
 
         self.yellow_circuit
-            .update_actuator_volumes(self.forward_cargo_door.actuator());
+            .update_system_actuator_volumes(self.forward_cargo_door.actuator());
 
         self.yellow_circuit
-            .update_actuator_volumes(self.aft_cargo_door.actuator());
+            .update_system_actuator_volumes(self.aft_cargo_door.actuator());
 
         self.yellow_circuit
-            .update_actuator_volumes(&mut self.nose_steering);
+            .update_system_actuator_volumes(&mut self.nose_steering);
 
-        self.yellow_circuit.update_actuator_volumes(
+        self.yellow_circuit.update_system_actuator_volumes(
             self.right_elevator
                 .actuator(ElevatorActuatorPosition::Inboard),
         );
 
         self.yellow_circuit
-            .update_actuator_volumes(self.rudder.actuator(RudderActuatorPosition::Yellow));
+            .update_system_actuator_volumes(self.rudder.actuator(RudderActuatorPosition::Yellow));
 
         self.yellow_circuit
-            .update_actuator_volumes(self.left_spoilers.actuator(1));
+            .update_system_actuator_volumes(self.left_spoilers.actuator(1));
         self.yellow_circuit
-            .update_actuator_volumes(self.left_spoilers.actuator(3));
+            .update_system_actuator_volumes(self.left_spoilers.actuator(3));
 
         self.yellow_circuit
-            .update_actuator_volumes(self.right_spoilers.actuator(1));
+            .update_system_actuator_volumes(self.right_spoilers.actuator(1));
         self.yellow_circuit
-            .update_actuator_volumes(self.right_spoilers.actuator(3));
+            .update_system_actuator_volumes(self.right_spoilers.actuator(3));
 
         self.yellow_circuit
-            .update_actuator_volumes(self.trim_assembly.right_motor());
+            .update_system_actuator_volumes(self.trim_assembly.right_motor());
     }
 
     fn update_blue_actuators_volume(&mut self) {
         self.blue_circuit
-            .update_actuator_volumes(self.slat_system.left_motor());
+            .update_system_actuator_volumes(self.slat_system.left_motor());
         self.blue_circuit
-            .update_actuator_volumes(&mut self.emergency_gen);
+            .update_system_actuator_volumes(&mut self.emergency_gen);
 
-        self.blue_circuit.update_actuator_volumes(
+        self.blue_circuit.update_system_actuator_volumes(
             self.left_aileron
                 .actuator(AileronActuatorPosition::Outboard),
         );
-        self.blue_circuit.update_actuator_volumes(
+        self.blue_circuit.update_system_actuator_volumes(
             self.right_aileron
                 .actuator(AileronActuatorPosition::Outboard),
         );
 
-        self.blue_circuit.update_actuator_volumes(
+        self.blue_circuit.update_system_actuator_volumes(
             self.left_elevator
                 .actuator(ElevatorActuatorPosition::Outboard),
         );
-        self.blue_circuit.update_actuator_volumes(
+        self.blue_circuit.update_system_actuator_volumes(
             self.right_elevator
                 .actuator(ElevatorActuatorPosition::Outboard),
         );
 
         self.blue_circuit
-            .update_actuator_volumes(self.rudder.actuator(RudderActuatorPosition::Blue));
+            .update_system_actuator_volumes(self.rudder.actuator(RudderActuatorPosition::Blue));
 
         self.blue_circuit
-            .update_actuator_volumes(self.left_spoilers.actuator(2));
+            .update_system_actuator_volumes(self.left_spoilers.actuator(2));
 
         self.blue_circuit
-            .update_actuator_volumes(self.right_spoilers.actuator(2));
+            .update_system_actuator_volumes(self.right_spoilers.actuator(2));
     }
 
     // All the core hydraulics updates that needs to be done at the slowest fixed step rate
@@ -2275,6 +2279,7 @@ impl A320Hydraulic {
             context,
             &mut vec![&mut self.engine_driven_pump_1],
             None::<&mut ElectricPump>,
+            None::<&mut ElectricPump>,
             Some(&self.power_transfer_unit),
             &self.green_circuit_controller,
             reservoir_pneumatics.green_reservoir_pressure(),
@@ -2290,6 +2295,7 @@ impl A320Hydraulic {
             context,
             &mut vec![&mut self.engine_driven_pump_2],
             Some(&mut self.yellow_electric_pump),
+            None::<&mut ElectricPump>,
             Some(&self.power_transfer_unit),
             &self.yellow_circuit_controller,
             reservoir_pneumatics.yellow_reservoir_pressure(),
@@ -2305,6 +2311,7 @@ impl A320Hydraulic {
             context,
             &mut vec![&mut self.blue_electric_pump],
             Some(&mut self.ram_air_turbine),
+            None::<&mut ElectricPump>,
             None,
             &self.blue_circuit_controller,
             reservoir_pneumatics.blue_reservoir_pressure(),