+++ .github/CHANGELOG.md
@@ -70,6 +70,7 @@
 1. [COND] Connect Air Conditioning System to Electrical System - @mjuhe (Miquel Juhe)
 1. [ADDON] Simbridge Integration - @lucky38i (Alex)
 1. [SD] Visual and functional improvements/fixes to the HYD SD page - @lukecologne (luke)
+1. [HYD] Added optional auxiliary hydraulic section in core hydraulic circuits - @Crocket63
 
 ## 0.8.0